/work to do on auth

+++ SneakerApp/Controllers/ReviewsController.cs
@@ -23,46 +23,85 @@ namespace SneakerApp.Controllers
             _roleManager = roleManager;
         }
 
-        // EDIT
+        // Editarea unui review asociat unui produs
+        // Se poate edita un review doar de catre utilizatorul
+        // care a postat review-ul respectiv
+
+        [Authorize(Roles = "User,Editor,Admin")]
         public IActionResult Edit(int id)
         {
             Review comm = db.Reviews.Find(id);
 
-            return View(comm);
+            //a se face cazul in care review-ul este NULL! - TBA
+            if (comm.UserId == _userManager.GetUserId(User) || User.IsInRole("Admin"))
+            {
+                return View(comm);
+            }
+            else
+            {
+                TempData["message"] = "Nu aveti dreptul sa editati review-ul";
+                TempData["messageType"] = "alert-danger";
+                return RedirectToAction("Index", "Products");
+            }
+
         }
 
 
         [HttpPost]
+        [Authorize(Roles = "User,Editor,Admin")]
         public IActionResult Edit(int id, Review requestReview)
         {
             Review comm = db.Reviews.Find(id);
 
-            if (ModelState.IsValid)
+            if (comm.UserId == _userManager.GetUserId(User) || User.IsInRole("Admin"))
             {
+                if (ModelState.IsValid)
+                {
 
-                comm.Content = requestReview.Content;
+                    comm.Content = requestReview.Content;
 
-                db.SaveChanges();
+                    db.SaveChanges();
 
-                return Redirect("/Products/Show/" + comm.ReviewId);
+                    return Redirect("/Products/Show/" + comm.ProductId);
+                }
+                else
+                {
+                    return View(requestReview);
+                }
             }
             else
             {
-                return View(requestReview);
+                TempData["message"] = "Nu aveti dreptul sa editati review-ul";
+                TempData["messageType"] = "alert-danger";
+                return RedirectToAction("Index", "Products");
             }
 
         }
 
         // DELETE
+        // Se poate sterge review-ul doar de catre userii cu rolul de Admin
+        // sau de catre utilizatorii cu rolul de User sau Editor (doar daca
+        // acel review a fost postat de catre acestia
         [HttpPost]
         [Authorize(Roles = "User,Editor,Admin")]
         public IActionResult Delete(int id)
         {
             Review comm = db.Reviews.Find(id);
-            db.Reviews.Remove(comm);
-            db.SaveChanges();
-            return Redirect("/Products/Show/" + comm.ProductId);
+
+            if (comm.UserId == _userManager.GetUserId(User) || User.IsInRole("Admin"))
+            {
+                db.Reviews.Remove(comm);
+                db.SaveChanges();
+                return Redirect("/Products/Show/" + comm.ProductId);
+            }
+            else
+            {
+                TempData["message"] = "Nu aveti dreptul sa editati review-ul";
+                TempData["messageType"] = "alert-danger";
+                return RedirectToAction("Index", "Products");
+            }
+
         }
-        
+
     }
 }
